fix bugs

+++ Z-Pawn/src/main/java/com/isahl/chess/pawn/endpoint/device/jpa/local/sqlite/service/QttStorageService.java
@@ -244,7 +244,7 @@ public class QttStorageService
     @Override
     public void cleanSession(long session)
     {
-        _SessionRepository.deleteById(session);
+        if(_SessionRepository.existsById(session)) {_SessionRepository.deleteById(session);}
     }
 
     @Cacheable(key = "#key",
